Add PersonEntity converter, EducationService

+++ repository/src/main/java/com/itbootcamp/starter/repository/EducationRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
  * Created by foooox on 21.7.17.
  */
 @Repository
-public interface EducationRepository extends CrudRepository<EducationEntity,Long>{
+public interface EducationRepository extends CrudRepository<EducationEntity,Integer>{
     List<EducationEntity> findByProfileId(int profileId);
     List<EducationEntity> findAllByProfile(ProfileEntity profileEntity);
 }
