+++ repository/src/main/java/com/itbootcamp/starter/repository/SkillRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * Created by Sergei on 7/21/2017.
  */
 @Repository
-public interface SkillRepository extends CrudRepository<SkillEntity, Long> {
+public interface SkillRepository extends CrudRepository<SkillEntity, Integer> {
 }
